+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-Nothing currently
+### Added 2021-02-15
+- UsernamePassword credentials are now supported through the use of a `username` tag on the secret
 
 ## [1.0.0] - 2019/02/04